requirements.txt
flask>=3.0.0
gunicorn>=21.0.0
pypdf>=4.0.0
pikepdf>=9.0.0
pdfminer.six>=20231228
pdf2image>=1.17.0
Pillow>=10.0.0
cryptography>=42.0.0
python-magic-bin>=0.4.14; sys_platform == "win32"
python-magic>=0.4.27; sys_platform != "win32"
SQLAlchemy>=2.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Limiter>=3.5.0
Flask-CORS>=4.0.0
python-json-logger>=2.0.7
requests>=2.31.0
pytest>=8.0.0
pytest-flask>=1.3.0
Werkzeug>=3.0.0
